@@ -9,10 +9,10 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^6.2.2",
-    "react-scripts": "^2.1.3",
+    "react-scripts": "^5.0.0",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
