@@ -1,9 +1,22 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import './styles.css';
+import MovieListing from '../MovieListing';
+import { apiKey } from '../../utils/movieApiKey';
 
 let Home = () => {
+
+    useEffect(() => {
+        let movieText = 'Harry';
+        let fechtApi = async () => {
+            let response = await fetch(`?apiKey=${ apiKey }&s=${ movieText }&type=movie`)
+            .catch(err => { console.log('Err:' + err) });
+        }
+    })
+
     return (
-        <h1>home</h1>
+        <main className='main-main'>
+            <MovieListing />
+        </main>
     )
 };
 
